# stickiness: add stickiness support