--- a/service_config.go
+++ b/service_config.go
@@ -71,6 +71,8 @@ type ServiceConfig struct {
 	// If there's no exact match, look for the default config for the service (/service/) and use the corresponding MethodConfig if it exists.
 	// Otherwise, the method has no MethodConfig to use.
 	Methods map[string]MethodConfig
+
+	stickinessKey *string
 }
 
 func parseDuration(s *string) (*time.Duration, error) {
@@ -145,6 +147,7 @@ type jsonMC struct {
 // TODO(lyuxuan): delete this struct after cleaning up old service config implementation.
 type jsonSC struct {
 	LoadBalancingPolicy *string
+	StickinessKey       *string
 	MethodConfig        *[]jsonMC
 }
 
@@ -158,6 +161,8 @@ func parseServiceConfig(js string) (ServiceConfig, error) {
 	sc := ServiceConfig{
 		LB:      rsc.LoadBalancingPolicy,
 		Methods: make(map[string]MethodConfig),
+
+		stickinessKey: rsc.StickinessKey,
 	}
 	if rsc.MethodConfig == nil {
 		return sc, nil