… Documentation/stickiness.md