@menghanl 2018-04-18
Benchmark result:
```
==== before ====
 
go run benchmark/benchmain/main.go -benchtime=10s -workloads=all -compression=on -maxConcurrentCalls=1000 -trace=off -reqSizeBytes=1 -respSizeBytes=1 -networkMode=Local
Unary-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_1B-respSize_1B-Compressor_true: 
50_Latency: 14071.5150 µs 	90_Latency: 16441.2350 µs 	99_Latency: 20031.4030 µs 	Avg latency: 14210.7170 µs 	Count: 703661 	13866 Bytes/op	173 Allocs/op	
Histogram (unit: µs)
Count: 703661  Min: 613.9  Max: 45838.0  Avg: 14210.72
------------------------------------------------------------
[     613.868000,      613.869000)       1    0.0%    0.0%  
[     613.869000,      613.875089)       0    0.0%    0.0%  
[     613.875089,      613.918257)       0    0.0%    0.0%  
[     613.918257,      614.224279)       0    0.0%    0.0%  
[     614.224279,      616.393728)       0    0.0%    0.0%  
[     616.393728,      631.773371)       0    0.0%    0.0%  
[     631.773371,      740.802589)       0    0.0%    0.0%  
[     740.802589,     1513.731516)       0    0.0%    0.0%  
[    1513.731516,     6993.172135)       1    0.0%    0.0%  
[    6993.172135,    45837.972000)  703658  100.0%  100.0%  ##########
[   45837.972000,             inf)       1    0.0%  100.0%  
 
Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_1B-respSize_1B-Compressor_true: 
50_Latency: 16236.1490 µs 	90_Latency: 49637.4370 µs 	99_Latency: 76499.9850 µs 	Avg latency: 23501.6690 µs 	Count: 425264 	41246 Bytes/op	40 Allocs/op	
Histogram (unit: µs)
Count: 425264  Min: 215.3  Max: 131301.7  Avg: 23501.67
------------------------------------------------------------
[      215.284000,       215.285000)       1    0.0%    0.0%  
[      215.285000,       215.291979)       0    0.0%    0.0%  
[      215.291979,       215.347665)       0    0.0%    0.0%  
[      215.347665,       215.791987)       0    0.0%    0.0%  
[      215.791987,       219.337250)       0    0.0%    0.0%  
[      219.337250,       247.625060)       0    0.0%    0.0%  
[      247.625060,       473.334737)       6    0.0%    0.0%  
[      473.334737,      2274.282144)      53    0.0%    0.0%  
[     2274.282144,     16644.120436)  221560   52.1%   52.1%  #####
[    16644.120436,    131301.690000)  203643   47.9%  100.0%  #####
[   131301.690000,              inf)       1    0.0%  100.0%  
 
==== after ====

go run benchmark/benchmain/main.go -benchtime=10s -workloads=all -compression=on -maxConcurrentCalls=1000 -trace=off -reqSizeBytes=1 -respSizeBytes=1 -networkMode=Local
Unary-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_1B-respSize_1B-Compressor_true: 
50_Latency: 13.8357 ms 	90_Latency: 15.9277 ms 	99_Latency: 19.8321 ms 	Avg latency: 13.9379 ms 	Count: 717441 13868 Bytes/op	174 Allocs/op	
Histogram (unit: ms)
Count: 717441  Min:   3.8  Max:  33.6  Avg: 13.94
------------------------------------------------------------
[       3.845331,        3.845332)       1    0.0%    0.0%  
[       3.845332,        3.845338)       0    0.0%    0.0%  
[       3.845338,        3.845377)       0    0.0%    0.0%  
[       3.845377,        3.845641)       0    0.0%    0.0%  
[       3.845641,        3.847429)       0    0.0%    0.0%  
[       3.847429,        3.859533)       0    0.0%    0.0%  
[       3.859533,        3.941452)       0    0.0%    0.0%  
[       3.941452,        4.495892)       0    0.0%    0.0%  
[       4.495892,        8.248425)     553    0.1%    0.1%  
[       8.248425,       33.646118)  716886   99.9%  100.0%  ##########
[      33.646118,             inf)       1    0.0%  100.0%  
 
Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_1000-reqSize_1B-respSize_1B-Compressor_true: 
50_Latency: 16425.5590 µs 	90_Latency: 48938.5340 µs 	99_Latency: 75172.7280 µs 	Avg latency: 23827.6740 µs 	Count: 419785 	42069 Bytes/op	40 Allocs/op	
Histogram (unit: µs)
Count: 419785  Min: 184.6  Max: 141370.9  Avg: 23827.67
------------------------------------------------------------
[      184.603000,       184.604000)       1    0.0%    0.0%  
[      184.604000,       184.611045)       0    0.0%    0.0%  
[      184.611045,       184.667724)       0    0.0%    0.0%  
[      184.667724,       185.123712)       0    0.0%    0.0%  
[      185.123712,       188.792190)       0    0.0%    0.0%  
[      188.792190,       218.305535)       2    0.0%    0.0%  
[      218.305535,       455.743926)       9    0.0%    0.0%  
[      455.743926,      2365.964217)      39    0.0%    0.0%  
[     2365.964217,     17733.915193)  232117   55.3%   55.3%  ######
[    17733.915193,    141370.921000)  187616   44.7%  100.0%  ####
[   141370.921000,              inf)       1    0.0%  100.0%
```