@MakMukhi 2018-04-18
*[picker_wrapper.go, line 73 at r4](https://beta.reviewable.io/reviews/grpc/grpc-go/1969#-LAP_4rk686Qmdf8Z1SA:-LAP_4rk686Qmdf8Z1SB:b-x6botz) ([raw file](https://github.com/grpc/grpc-go/blob/e8b72c1cde17b08dc56c3f30891f90d3fce8f569/picker_wrapper.go#L73)):*
> ```Go
> 
> func (bp *pickerWrapper) updateStickinessMDKey(newKey string) {
> 	if oldKey, _ := bp.stickinessMDKey.Load().(string); oldKey != newKey {
> ```

How about `atomic.Swap` instead of `atomic.Load` ?

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/grpc/grpc-go/1969)*
<!-- Sent from Reviewable.io -->