@MakMukhi 2018-04-18
*[picker_wrapper.go, line 80 at r4](https://beta.reviewable.io/reviews/grpc/grpc-go/1969#-LAP_TSV1EFd_fRK-q9m:-LAP_TSV1EFd_fRK-q9n:b-ya81ng) ([raw file](https://github.com/grpc/grpc-go/blob/e8b72c1cde17b08dc56c3f30891f90d3fce8f569/picker_wrapper.go#L80)):*
> ```Go
> 
> func (bp *pickerWrapper) clearStickinessState() {
> 	if oldKey, _ := bp.stickinessMDKey.Load().(string); oldKey != "" {
> ```

Same as above

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/grpc/grpc-go/1969)*
<!-- Sent from Reviewable.io -->