@MakMukhi 2018-04-18
*[stickiness_test.go, line 172 at r4](https://beta.reviewable.io/reviews/grpc/grpc-go/1969#-LAPcmXsB9Glns4e6hbX:-LAPcmXsB9Glns4e6hbY:b-eskks5) ([raw file](https://github.com/grpc/grpc-go/blob/e8b72c1cde17b08dc56c3f30891f90d3fce8f569/stickiness_test.go#L172)):*
> ```Go
> 	}
> 
> 	if picked[0] != 0 && picked[1] != 0 {
> ```

How about: ` (picked[0] != 0 && picked[1] == 0) || (picked[0] == 0 && picked[1] !=0) ` ?

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/grpc/grpc-go/1969)*
<!-- Sent from Reviewable.io -->